backend-python/requirements.txt
fastapi>=0.116,<1.0
uvicorn[standard]>=0.35,<1.0
supabase>=2.15,<3.0
httpx>=0.28,<1.0
python-dotenv>=1.1,<2.0
email-validator>=2.2,<3.0
